requirements.txt

timm==0.6.11
opencv-python==4.6.0.66
diffdist==0.1
nltk>=3.6.2
einops>=0.3.0
omegaconf==2.1.1
open-clip-torch==2.0.2
stable-diffusion-sdkit==2.1.3

importlib-metadata==4.11.3
triton==2.0.0.post1